requirements.txt
asgiref==3.2.3
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.3
django-extensions==2.2.8
django-storages==1.9.1
entrypoints==0.3
enums==0.0.2
google==2.0.3
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-core==1.3.0
google-cloud-speech==1.3.2
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.1
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
moviepy==1.0.1
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
pandocfilters==1.4.2
parso==0.6.2
pickleshare==0.7.5
Pillow==7.0.0
proglog==0.1.9
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydub==0.23.1
Pygments==2.5.2
pyrsistent==0.15.7
python-dateutil==2.8.1
pytube3==9.6.4
pytz==2019.3
pywinpty==0.5.7
pyzmq==19.0.0
qtconsole==4.7.0
QtPy==1.9.0
requests==2.23.0
rsa==4.0
Send2Trash==1.5.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
tqdm==4.43.0
traitlets==4.3.3
typing-extensions==3.7.4.1
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.0.0
